// SPA/src/app/modules/dal/enums/EModalName.ts
export enum EModalName {
    CreateUserForm = 'create-user-form',
    UpdateUserForm = 'update-user-form',
    ConfirmDeletePopup = 'confirm-delete-popup',

    None = 'none',
    AddCamera = 'add-camera',
    EditCamera = 'edit-camera',
    LiveCamera = 'live-camera',

    EditUser = 'edit-user',
    AddLocation = 'add-location',
    EditLocation = 'edit-location'
}